Add a commit hook that checks for whyCommit use
Addresses
---------
The need for an automated check to ensure that a reason is provided.

Why
---
The commit hook is written as a bash script because its very likely that git is used in a bash shell.
The "Why\n---" format is somewhat arbitrary, but not a bad one. Finally the # for the small commit seemed the least surprising one, especially since the text is rewritten without it.
Gotchas: bash's regex operator slurps up all newlines in the collection. This is a bug that needs to be fixed.